Server/requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
feedparser==5.1.3
FlexGet==1.2
html5lib==0.999
Jinja2==2.7.2
jsonschema==2.3.0
MarkupSafe==0.19
progressbar==2.2
pyDatalog==0.14.5
pynzb==0.1.0
PyPubSub==3.3.0
PyRSS2Gen==1.1
python-dateutil==2.1
python-tvrage==0.4.1
PyYAML==3.11
requests==2.2.1
rpyc==3.2.3
six==1.6.1
SQLAlchemy==0.9.4
tmdb3==0.6.17
virtualenv==1.11.4
